apply v2 layout if it's legal exercise else apply the original layout (v1)
